\name{rkafka-package}
\alias{rkafka-package}
\alias{rkafka}
\docType{package}
\title{
\packageTitle{rkafka}
}
\description{
\packageDescription{rkafka}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rkafka}
\packageIndices{rkafka}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rkafka}

Maintainer: \packageMaintainer{rkafka}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
print(kafka_version())
}
